requirements.txt
google-api-python-client==2.95.0
Flask==2.3.2
DateTime==5.2
regex==2023.6.3
python-dotenv== 1.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.6
click==8.1.6
blinker==1.6.2
itsdangerous==2.1.2
google-api-core==2.11.1
google-auth==2.22.0
google-auth-httplib2==0.1.0
httplib2==0.22.0
uritemplate==4.1.1
googleapis-common-protos==1.59.1
protobuf==4.23.4
requests==2.31.0
cachetools====5.3.1
pyasn1-modules====0.3.0
rsa====4.9
six==1.16.0
urllib3==1.26.16
pyparsing==3.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pyasn1==0.5.0
pytz==2023.3
zope.interface==6.0
setuptools==68.0.0
gunicorn==21.2.0
packaging==23.1
colorama==0.4.6
pip==23.1.2
wheel==0.40.0